% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroke_prediction_save_service.R
\name{stroke_prediction_save_service}
\alias{stroke_prediction_save_service}
\title{Shanghai stroke prediction and save service system
transfer data from shanghai fifth hospital to huashan}
\usage{
stroke_prediction_save_service(...)
}
\arguments{
\item{...}{one or more xlsx file names}
}
\value{
write out csv file
}
\description{
Shanghai stroke prediction and save service system
transfer data from shanghai fifth hospital to huashan
}
\examples{
\donttest{
stroke_prediction_save_service('1.xlsx','2.xlsx')
}
}
